Add files via upload
This Matlab script computes the Aldaz-Kouncev-Render (AKR) and Bernstein approximating operators of a function in 1d and 2d. 
The class of functions on  which the AKR- operators is better than the Bernstein one are described in the paper 
A. Acu, S. De Marchi and Ioan Rasa: Aldaz-Kounchev-Render operators and their approximation properties. 

For any further questions please write to stefano.demarchi@unipd.it