Add files via upload
This script computes the Aldaz-Kouncev-Render and  Bernstein approximating operators
of a function on a triangle. The functions considered are exp(x^2y^2) and exp(-x-y). 
More details are in the paper 
A. Acu, S. De Marchi and I. Rasa: Aldaz-Kounchev-Render operators on simplices.

For any further inquires, please email to stefano.demarchi@unipd.it